requirements.txt
h5py==2.9.0
numpy==1.15.4
pandas==0.23.4
pillow==5.4.1
plyfile==0.7
scikit-image==0.15.0
scipy==1.2.1
visdom==0.1.7
